requirements.txt
# CivicNav - Azure AI Engineering Lab
# Core dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Azure SDKs
azure-identity>=1.17.0
azure-search-documents>=11.6.0
openai>=1.40.0

# MCP Server
mcp>=1.0.0

# Utilities
python-multipart>=0.0.9
httpx>=0.27.0
